feat: allow decimal scores (e.g. 7.6, 5.5) in all evaluation forms
- Remove .int() from Zod scoreField in episodic-evaluations and episodes validations
- Remove Number.isInteger() checks from episode and call-report revision API routes
- Remove .int() from initial-assessments API schema
- ScoreCard: parseInt → parseFloat, add step="0.1", widen input w-16→w-20, add placeholder
- DB migration: ALTER score columns from INT to NUMERIC(4,1) in episodic_evaluations,
  evaluator_forms, and episodes tables (drop/re-add CHECK constraints)

+++ app/api/initial-assessments/route.ts
@@ -17,7 +17,7 @@ export const dynamic = 'force-dynamic';
 const upsertSchema = z.object({
     entity_type: z.enum(["call_report", "episode"]),
     entity_id: z.string().uuid(),
-    score: z.number().int().min(1).max(10),
+    score: z.number().min(1).max(10),
     comment: z.string().max(1000).optional().nullable(),
 });
